backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitarray==3.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
cytoolz==1.0.1
deprecation==2.1.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
frozendict==2.3.10
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hexbytes==1.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
moralis==0.1.49
multidict==6.1.0
packaging==24.2
parsimonious==0.10.0
postgrest==0.19.3
propcache==0.3.0
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
toolz==1.0.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
